Split base/__init__.py into four files
